Remove need to sudo.

+++ install.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+if [ -e $SUDO_USER ]; then
+    sudo -E ./install.sh
+    exit
+fi
+
 apt update
 apt install -y \
     build-essential gcc autotools-dev automake libncurses5-dev libncursesw5-dev libevent-dev xclip bison pkg-config
